fix(rag_web_demo): Fixed test_api_connection gr.Error unable to render issues and encapsulate the http api in a single package
